Michigan State University & Western Michigan University Peer Institutions Datasets

The Michigan State University (MSU) dataset contains institutions within the BigTen, the PAC-12, are Landgrant institutions, are public Association of American Universities (AAU), Mid-American Conference (aka the MAC), institutions of note within the state of Michigan, and other institutions of interest.

The Western Michigan University (WMU) dataset is the WMU's Institutional Research's Peer Institutions list. The WMU Peer Institutions are based on several key parameters from the Carnegie Classification of Institutions of Higher Education.

The information common to each dataset:

INSTNM: Institution name

name.short: Short name for plots and reports

UNITID: IPEDS unit ID

OPEID:

OPEID6:

CITY: City

STABBR: State

ZIP: Zip code

index: Index to aid reordering

colour.primary: Institution's primary colour

colour.secondary: Institution's secondary colour

The institution's primary and secondary colours were, in most cases, from the institution's branding page/website.

The MSU peers dataset also includes the following columns to aid the construction of subsets

BigTen.tf: BigTen institution

PAC12.tf: PAC-12 institution

MAC.tf: MAC institution

Public.tf: Public institution

Landgrant.tf: Landgrant institution (based on List of land-grant universities Wikipedia entry)

AAU.tf: Association of American Universities (based on AAU provided list)

MIinsts.tf: Institutions within the state of Michigan

OIOI.tf: Other institutions of interest

MSUpeer.tf: MSU peer institution

The selection of the WMU peer institutions used the 2015 Carnegie Classification categories:

Control: Public

Level: Four-year or above

Basic Classification: Doctoral universities, higher research activity

Undergraduate Instructional Program: Professions plus arts & sciences, high graduate coexistence

Graduate Instructional Program: Research doctoral, comprehensive programs, no medical/veterinary school

Enrollment Profile: High undergraduate

Undergraduate Profile: Four-year, full-time, selective, higher transfer-in

Size and Setting: Four-year, large, primarily residential

Total enrollment; total research expenditures; and total doctoral degrees awarded

The WMU peer instititions:

Ball State University (Muncie, IN)

East Carolina University (Greenville, NC)

Kent State University at Kent (Kent, OH)

Northern Arizona University (Flagstaff, AZ)

Northern Illinois University (Dekalb, IL)

Ohio University - Main Campus (Athens, OH)

Oklahoma State University - Main Campus (Stillwater, OK)

Portland State University (Portland, OR)

SUNY at Binghamton (Vestal, NY)

The University of Alabama (Tuscaloosa, AL)

University of Memphis (Memphis, TN)

University of Nevada - Las Vegas (Las Vegas, NV)

University of North Carolina at Greensboro (Greensboro, NC)

University of North Dakota (Grand Forks, ND)

University of Southern Mississippi (Hattiesburg, MS)

Usage

MSUpeers

WMUpeers

Format

A tibble of institutions.

An object of class tbl_df (inherits from tbl, data.frame) with 16 rows and 11 columns.

Source

Data sources:

Landgrant institutions (based on List of land-grant universities Wikipedia entry)

Association of American Universities (based on AAU provided list)

WMU's Institutional Research's Peer Institutions

Carnegie Classification of Institutions of Higher Education

Functions

WMUpeers: Western Michigan Peer Institutions dataset

Author

Emilio Xavier Esposito emilio@msu.edu (https://github.com/emilioxavier)