Longitude-Latitude to Region

Identify the region (and subregion) of a geospatial location position using longitude and latitude.

Usage

lonlat2region(longitude, latitude, gadm.mapfile)

Arguments

longitude

numeric value indicating the longitude

latitude

numeric value indicating the latitude

gadm.mapfile

string with the path and file name of the GADM world mapfile. Any mapfile can be used, but for datasets containing geospatial location data from anywhere on the planet, it is advantageous to use the entire world map. Obtain the mapfile of interest from https://gadm.org. The world GADM data is available from https://biogeo.ucdavis.edu/data/gadm3.6/gadm36_gpkg.zip. Ensure the GeoPackage format is downloaded. The compressed complete world database is approximately 1.6 GBs and the unzipped file is 3.81 GBs. For example, ~/mapfiles/gadm36_levels.gpkg

Individual country maps are available from https://gadm.org/maps.html.

Value

tibble with the country.code, country.name, region, and subregion.

Details

Based on the provided longitude and latitude, the corresponding country, region, and subregion are identified. Positions outside of a country return NAs for the country.code, country.name, region, and subregion. Instances where NA is provided for longitude and latitude, also returns NA values.

When using the world GADM database, loading the rather large world database takes time and, thus, the function appears to be slow.

This function is based on the "Latitude Longitude Coordinates to State Code in R" stackoverflow response by Josh O'Brien.

Author

Emilio Xavier Esposito emilio@msu.edu (https://github.com/emilioxavier)

Examples

if (FALSE) {
mapfile.world <- "~/mapfiles/gadm36_levels.gpkg"
lonlat2regionfunction(longitude, latitude, gadm.mapfile=mapfile.world)
}