Construct Donut Plot Data

add the description of making the data

Usage

make.donut.data(
  data,
  col.oi,
  facetBy = NULL,
  category.order = "count",
  levels.rev = FALSE,
  col.count = NULL,
  r.inner = 4,
  r.outer = 6
)

Arguments

data

tibble (or data.frame) with the column of interest. NOTE: Do NOT use counted data.

col.oi

string with column of interest. Only provide ONE column name.

facetBy

string indicating the column to group data by; for when you want to facet your donut plots via ggplot2::facet_wrap(); see make.donut.plot().

category.order

string indicating if you want the data to be ordered by "count" in decreasing order or by "category" in alphabetical order; default: "count".

levels.rev

logical indicating if the order of the categories should be reversed.

col.count

string with the column containing "counts" for each "category." This parameter is required when the count (or total) for each row was pre-calculated and allows for the creation of donut data when raw data is not available and one only has the summarised values.

r.inner

numeric value defining the inner radius of the donut; default: 4

r.outer

numeric value defining the outer radius of the donut; default: 6

Value

tibble with the

Author

Emilio Xavier Esposito emilio@msu.edu (https://github.com/emilioxavier)

Examples

if (FALSE) {
  donut.DATA <- make.donut.data(data, col.oi, facetBy=NULL,
                                category.order="count", levels.rev=FALSE,
                                r.inner=4, r.outer=6)
}